refactor: standalone non-optional RendererSettings
RendererSettings drops the Theme inheritance and re-declares the field block with concrete types (None survives only where it is a resolved value: background_color, commit_row_band_color); the resolved-pixel accessors and branch_label_side move with it. Lifts the ty excludes so the whole library type-checks, amends architecture invariant #8, and adds a field-parity meta-test so the mirrored block cannot drift. Rendered output byte-identical.

+++ gitsvg/render/_primitives/commit_label.py
@@ -110,13 +110,13 @@ def draw_commit_label(d: draw.Drawing, commit: LayoutCommit, canvas: RenderCanva
         )
 
 
-def _build_lines(commit: LayoutCommit, theme: RendererSettings) -> list[tuple[str, int, str, str]]:
+def _build_lines(commit: LayoutCommit, theme: RendererSettings) -> list[tuple[str, float, str, str]]:
     r"""Return the line stack as `(text, font_size, color, font_weight)` tuples.
 
     Order: each `msg` split on `"\\n"` becomes a primary line; the
     optional `hash` follows as a single secondary line.
     """
-    lines: list[tuple[str, int, str, str]] = []
+    lines: list[tuple[str, float, str, str]] = []
     msg_weight = "700" if commit.highlight else "400"
     if commit.msg is not None:
         for msg_line in commit.msg.split("\n"):
